feat: refactoring using replace extract class
