+++ retrofit/src/main/java/retrofit2/ServiceMethod.java
@@ -15,7 +15,7 @@
  */
 package retrofit2;
 
-import static retrofit2.Utils.methodError;
+import static retrofit2.ErrorsUtils.methodError;
 
 import java.lang.reflect.Method;
 import java.lang.reflect.Type;
